mde the changes requested
